refactor contruct method for RMS and Kurtosis

+++ src/main/java/com/vape/data/streaming/service/DataPointConsumer.java
@@ -36,7 +36,7 @@ public class DataPointConsumer {
         log.info(String.format(LOG_MSG, message));
         SensorDataPointModel dataPointModel = mapper.toObject(message, SensorDataPointModel.class);
         BigDecimal computedRMS = dspEngineService.computeRMS(dataPointModel);
-        RMSDataPointModel rmsDataPointModel = constructRMSDataPointModel(dataPointModel, computedRMS.doubleValue());
+        RMSDataPointModel rmsDataPointModel = constructRMSDataPointModel(dataPointModel, computedRMS);
         dataPointProducer.publishRMS(rmsDataPointModel);
     }
 
@@ -45,23 +45,23 @@ public class DataPointConsumer {
         log.info(String.format(LOG_MSG, message));
         SensorDataPointModel dataPointModel = mapper.toObject(message, SensorDataPointModel.class);
         BigDecimal computedKurtosis = dspEngineService.computeKurtosis(dataPointModel);
-        KurtosisDataPointModel kurtosisDataPointModel = constructKurtosisDataPointModel(dataPointModel, computedKurtosis.doubleValue());
+        KurtosisDataPointModel kurtosisDataPointModel = constructKurtosisDataPointModel(dataPointModel, computedKurtosis);
         dataPointProducer.publishKurtosis(kurtosisDataPointModel);
     }
 
-    KurtosisDataPointModel constructKurtosisDataPointModel(SensorDataPointModel dataPointModel, Double computedData) {
+    KurtosisDataPointModel constructKurtosisDataPointModel(SensorDataPointModel dataPointModel, BigDecimal computedData) {
         return KurtosisDataPointModel.builder()
                 .sensorDataPointId(dataPointModel.getId())
                 .timestamp(LocalDateTime.now().toString())
-                .data(computedData)
+                .data(computedData.doubleValue())
                 .build();
     }
 
-    RMSDataPointModel constructRMSDataPointModel(SensorDataPointModel dataPointModel, Double computedData) {
+    RMSDataPointModel constructRMSDataPointModel(SensorDataPointModel dataPointModel, BigDecimal computedData) {
         return RMSDataPointModel.builder()
                 .sensorDataPointId(dataPointModel.getId())
                 .timestamp(LocalDateTime.now().toString())
-                .data(computedData)
+                .data(computedData.doubleValue())
                 .build();
     }
 }
